requirements.txt
python-dotenv
pandas
numpy
pymongo
Certifi
pymongo[srv]==3.11
scikit-learn
dill
PyYAML
# -e .